Make a final commit after sorting the meta file
When retrying compaction that generated a large meta file we would take
a considerable amount of time seeking through the data generated by the
id merge sort. This would take so long that the fd for the data file
would be closed by couch_file (because it's not monitored yet) which
would then blow up when trying to read from it once we read the meta
file header.

This patch just sorts the meta file in one step, rewrites #comp_header{}
entries for both files and then does the streaming copy of ids to the
data file. This way when we retry the meta file will have a header right
at the end of the file so opening won't take a long time.