Remove stale optimization for compaction commits
Before the new compactor we hacked the fsync options to only commit once
before the header was written instead of before and after. This stale
override then lead to an issue where the final commit_data call didn't
fsync after a #db_header{} was written leaving a #comp_header{} as the
last valid header which lead to errors about corruption when opening the
database later on.