Add delegated authentication module
This module allows a server administrator to create a cookie that, when presented by a client will authenticate that client and authorize them with the username and roles chosen by the administrator.

Setup
=====

[httpd_global_handlers]
_delegated_auth = {couch_httpd_delegated, handle_delegated_auth_req}

[httpd]
authentication_handlers = {couch_httpd_delegated, delegated_authentication_handler}

Usage
=====

A server administrator generates a client cookie;

> curl -X POST -u admin -Hcontent-type:application/json localhost:5984/_delegated_auth -d '{"name":"foo","roles":["bar", "baz"]}'
< {"ok":true,"cookie":"DelegatedAuth=Zm9vOmJheixiYXI6NEY0RTcwODU684aMS3jtFAjKIrUHlTgal37EtBo; Version=1; Expires=Wed, 29 Feb 2012 19:37:57 GMT; Max-Age=3600; Path=/"}

The mechanism by which the administration passes the value of the cookie property is not specified.

A client uses the cookie as follows;

> curl -H"cookie:DelegatedAuth=Zm9vOmJheixiYXI6NEY0RTcwODU684aMS3jtFAjKIrUHlTgal37EtBo; Version=1; Expires=Wed, 29 Feb 2012 19:37:57 GMT; Max-Age=3600; Path=/" localhost:5984/_session
< {"ok":true,"userCtx":{"name":"foo","roles":["baz","bar"]},"info":{"authentication_db":"_users","authentication_handlers":["delegated","cookie","default"],"authenticated":"delegated"}}