Force db headers to disk after compaction
There's a race condition that can happen if we die just after compaction
has finished writing its last compaction header but before
couch_db_udpater has swapped out the files. If we resume the compaction
we hit an optimization in couch_db_updater:commit_data/2 that avoids
writing a header if it hasn't changed.

This is just a simple patch to force the db header to disk after
compaction has finished.

BugzId: 14936